Added the capability to draw multiple diagrams
We were using singleton for accessing diagramManager. I removed
singleton so we can have a diagramManager for each diagram. This broke
the properties window, but since the properties window has not been
fully implemented and needs to correlate with each diagram window I did
not see it as a huge problem.