Implemented all relationships
AggregationRelationship is currently hard-coded within diagramManager
for createRelationship. Someone needs to implement the mechanism for
changing/creating different relationships. It'd be nice if someone could
implement the arrowhead for the dependency relationship since it looks
the same as realization relationship because I didn't want to code an
arrowhead just yet.