Reading/Writing XML broken
With the recent changes reading/writing XML has been broken. I've looked
over most of it and the last part that needs to be fixed is writing the
unique ID of the class to file and reading it back in to set the data on
the class as opposed to a new one being generated. This will allow
relationships to correctly reference the source and destination classes.
Relationship drawing has also been broken with the recent changes.