Added package name to class
Package name is now written to file and read from file for each class. I
also added package name to the class properties form, which can also be
edited. I'd like to change package name and diagram name so it doesn't
ask the user for each name, but instead uses the name saved to file.